pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ru-it-audit-notebook"
version = "0.1.0"
description = "Interactive IT audit report templates for Russian contexts"
readme = "README.md"
authors = [{name = "run-as-daemon.ru", email = "dev@run-as-daemon.ru"}]
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
dependencies = [
  "typer>=0.9",
  "pyyaml>=6.0",
  "jinja2>=3.1",
  "pandas>=2.1",
  "papermill>=2.4",
  "nbformat>=5.9",
  "nbconvert>=7.9",
  "click>=8.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "pytest-cov>=4.1",
  "mypy>=1.6",
  "ruff>=0.1",
  "bandit>=1.7",
  "pip-audit>=2.6",
  "safety>=3.0",
  "yamllint>=1.32",
  "nbqa>=1.7",
  "types-pyyaml",
  "types-click",
]

[project.scripts]
ru-it-audit = "it_audit_report.cli.main:app"

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_unused_configs = true
warn_unused_ignores = true
check_untyped_defs = true
ignore_missing_imports = true
